# openapi.v1.yaml
openapi: 3.1.1
info:
  title: TNPSC Book RAG API
  version: 1.0.0
  summary: Frontend contract for the English-first TNPSC textbook RAG MVP.
  description: |
    This checked-in contract contains implemented and planned v1 operations so the frontend can
    generate types and mocks before every backend route exists. `x-implementation-status` records
    current delivery state. FastAPI's generated OpenAPI must conform to the implemented subset.
servers:
  - url: "{origin}"
    variables:
      origin:
        default: http://127.0.0.1:8000
        description: Deployment-specific API origin.
...
paths:
...
  /v1/sources/{chunk_id}:
    parameters:
      - $ref: "#/components/parameters/ChunkId"
    get:
      tags: [retrieval]
      operationId: getCitationSource
      summary: Get a stable evidence source for a citation.
      x-implementation-status: planned
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Citation evidence.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/search:
    post:
      tags: [retrieval]
      operationId: searchTextbooks
      summary: Perform standalone local semantic search.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Ranked textbook evidence.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/answers:
    post:
      tags: [answers]
      operationId: createAnswer
      summary: Retrieve evidence and generate a validated answer.
      description: |
        Missing or JSON Accept headers select the synchronous JSON representation. An explicit
        `Accept: text/event-stream` selects the SSE representation. The terminal SSE completion
        carries the same AnswerResponse schema as the JSON representation.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      x-sse-events:
        answer.started:
          $ref: "#/components/schemas/AnswerStartedEvent"
        answer.progress:
          $ref: "#/components/schemas/AnswerProgressEvent"
        answer.delta:
          $ref: "#/components/schemas/AnswerDeltaEvent"
        answer.reset:
          $ref: "#/components/schemas/AnswerResetEvent"
        answer.completed:
          $ref: "#/components/schemas/AnswerResponse"
        answer.failed:
          $ref: "#/components/schemas/Problem"
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: Accept
          in: header
          required: false
          schema:
            type: string
            enum: [application/json, text/event-stream, "*/*"]
            default: application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerRequest"
      responses:
        "200":
          description: Synchronous answer or SSE answer event stream.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResponse"
            text/event-stream:
              schema:
                type: string
                description: Named SSE events described by x-sse-events.
        "422":
          $ref: "#/components/responses/Unprocessable"
        "429":
          $ref: "#/components/responses/CapacityExceeded"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token issued by the configured OpenID Connect provider.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        $ref: "#/components/schemas/RequestIdentifier"
...
    ChunkId:
      name: chunk_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    RequestId:
      required: true
      schema:
        $ref: "#/components/schemas/RequestIdentifier"
...
  responses:
    NotFound:
      description: Resource was not found.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Unprocessable:
      description: Validation, cursor, or document-content error.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    CapacityExceeded:
      description: A bounded local or provider capacity limit was reached.
      headers:
        Retry-After:
          schema:
            oneOf:
              - type: integer
                minimum: 0
              - type: string
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ServiceUnavailable:
      description: A required service is temporarily unavailable.
      headers:
        Retry-After:
          schema:
            oneOf:
              - type: integer
                minimum: 0
              - type: string
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    RequestIdentifier:
      type: string
      minLength: 1
      maxLength: 200
      description: Accepted or generated safe printable correlation identifier.
    TextbookStandard:
      type: integer
      enum: [6, 7, 8, 9, 10]
...
    AssetType:
      type: string
      enum: [image, diagram, map, photograph, figure, unknown]
    AltTextSource:
      type: string
      enum: [caption, manual, unavailable, not_applicable]
    ChunkContentType:
      type: string
      enum: [prose, heading, list, table, caption, mixed]
    AnswerMode:
      type: string
      enum: [textbook_only, textbook_plus_general]
    ResponseLength:
      type: string
      enum: [short, medium, long]
    TextbookAnswerStatus:
      type: string
      enum: [answered, insufficient_evidence]
...
    ValidationFieldError:
      type: object
      required: [field, message, code]
      properties:
        field:
          type: string
          description: >-
            Stable path such as body.filters.standards.0, query.limit, path.book_id,
            header.Idempotency-Key, or form.file.
        message: {type: string}
        code: {type: string}
      additionalProperties: false
    Problem:
      type: object
      required: [type, title, status, detail, instance, code, request_id, errors]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        code: {type: string}
        request_id:
          $ref: "#/components/schemas/RequestIdentifier"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationFieldError"
      additionalProperties: false
...
    AssetReference:
      type: object
      required:
        - id
        - asset_type
        - caption
        - alt_text
        - alt_text_source
        - is_decorative
        - pixel_width
        - pixel_height
        - content_url
        - thumbnail_url
        - thumbnail_pixel_width
        - thumbnail_pixel_height
      properties:
        id: {type: string, format: uuid}
        asset_type:
          $ref: "#/components/schemas/AssetType"
        caption: {type: [string, "null"]}
        alt_text: {type: [string, "null"]}
        alt_text_source:
          $ref: "#/components/schemas/AltTextSource"
        is_decorative: {type: boolean}
        pixel_width: {type: [integer, "null"], minimum: 1}
        pixel_height: {type: [integer, "null"], minimum: 1}
        content_url: {type: string, pattern: "^/v1/assets/[^/]+/content$"}
        thumbnail_url:
          type: [string, "null"]
          pattern: "^/v1/assets/[^/]+/thumbnail$"
        thumbnail_pixel_width: {type: [integer, "null"], minimum: 1}
        thumbnail_pixel_height: {type: [integer, "null"], minimum: 1}
      description: >-
        Decorative assets have null alt_text and alt_text_source=not_applicable. Non-decorative
        assets use caption or manual alt text when available; unavailable is explicit. Thumbnail
        dimensions and URL are all null together when no raster thumbnail can be generated.
...
    Evidence:
      type: object
      required:
        - chunk_id
        - page_id
        - document_id
        - book_id
        - book_title
        - edition
        - standard
        - subject
        - pdf_page_index
        - printed_page_label
        - section_path
        - content_type
        - text
        - assets
        - source_url
      properties:
        chunk_id: {type: string, format: uuid}
        page_id: {type: string, format: uuid}
        document_id: {type: string, format: uuid}
        book_id: {type: string, format: uuid}
        book_title: {type: string}
        edition: {type: [string, "null"]}
        standard:
          $ref: "#/components/schemas/TextbookStandard"
        subject: {type: string}
        pdf_page_index: {type: integer, minimum: 0}
        printed_page_label: {type: [string, "null"]}
        section_path:
          type: array
          items: {type: string}
        content_type:
          $ref: "#/components/schemas/ChunkContentType"
        text: {type: string}
        assets:
          type: array
          items:
            $ref: "#/components/schemas/AssetReference"
        source_url: {type: string, pattern: "^/v1/sources/[^/]+$"}
...
    SearchFilters:
      type: object
      properties:
        standards:
          type: array
          maxItems: 5
          uniqueItems: true
          default: []
          items:
            $ref: "#/components/schemas/TextbookStandard"
        subjects:
          type: array
          maxItems: 20
          uniqueItems: true
          default: []
          items: {type: string, minLength: 1, maxLength: 200}
        book_ids:
          type: array
          maxItems: 50
          uniqueItems: true
          default: []
          items: {type: string, format: uuid}
        document_ids:
          type: array
          maxItems: 50
          uniqueItems: true
          default: []
          items: {type: string, format: uuid}
      additionalProperties: false
    SearchRequest:
      type: object
      required: [query]
      properties:
        query: {type: string, minLength: 1, maxLength: 1000}
        filters:
          $ref: "#/components/schemas/SearchFilters"
        top_k: {type: integer, minimum: 1, maximum: 50, default: 10}
      additionalProperties: false
    SearchHit:
      type: object
      required: [rank, score, score_kind, evidence]
      properties:
        rank: {type: integer, minimum: 1}
        score: {type: number}
        score_kind: {type: string, const: cosine_similarity}
        evidence:
          $ref: "#/components/schemas/Evidence"
      additionalProperties: false
    SearchResponse:
      type: object
      required: [query, top_k, filters, results]
      properties:
        query: {type: string}
        top_k: {type: integer, minimum: 1, maximum: 50}
        filters:
          $ref: "#/components/schemas/SearchFilters"
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchHit"
      additionalProperties: false
...
    InlineNode:
      oneOf:
        - $ref: "#/components/schemas/TextNode"
        - $ref: "#/components/schemas/CitationNode"
      discriminator:
        propertyName: type
        mapping:
          text: "#/components/schemas/TextNode"
          citation: "#/components/schemas/CitationNode"
    ParagraphBlock:
      type: object
      required: [type, nodes]
      properties:
        type: {type: string, const: paragraph}
        nodes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/InlineNode"
      additionalProperties: false
    BulletListBlock:
      type: object
      required: [type, items]
      properties:
        type: {type: string, const: bullet_list}
        items:
          type: array
          minItems: 1
          items:
            type: array
            minItems: 1
            items:
              $ref: "#/components/schemas/InlineNode"
      additionalProperties: false
    AnswerBlock:
      oneOf:
        - $ref: "#/components/schemas/ParagraphBlock"
        - $ref: "#/components/schemas/BulletListBlock"
      discriminator:
        propertyName: type
        mapping:
          paragraph: "#/components/schemas/ParagraphBlock"
          bullet_list: "#/components/schemas/BulletListBlock"
    SupplementaryParagraphBlock:
      type: object
      required: [type, nodes]
      properties:
        type: {type: string, const: paragraph}
        nodes:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/TextNode"
      additionalProperties: false
    SupplementaryBulletListBlock:
      type: object
      required: [type, items]
      properties:
        type: {type: string, const: bullet_list}
        items:
          type: array
          minItems: 1
          items:
            type: array
            minItems: 1
            items:
              $ref: "#/components/schemas/TextNode"
      additionalProperties: false
    SupplementaryBlock:
      oneOf:
        - $ref: "#/components/schemas/SupplementaryParagraphBlock"
        - $ref: "#/components/schemas/SupplementaryBulletListBlock"
      discriminator:
        propertyName: type
        mapping:
          paragraph: "#/components/schemas/SupplementaryParagraphBlock"
          bullet_list: "#/components/schemas/SupplementaryBulletListBlock"
    EvidenceCitation:
      allOf:
        - $ref: "#/components/schemas/Evidence"
        - type: object
          required: [citation_id]
          properties:
            citation_id: {type: string, pattern: "^T[1-9][0-9]*$"}
    TextbookAnswer:
      type: object
      required: [status, blocks, citations]
      properties:
        status:
          $ref: "#/components/schemas/TextbookAnswerStatus"
        blocks:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/AnswerBlock"
        citations:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceCitation"
      description: >-
        answered responses have at least one block and every citation node resolves to exactly one
        citation. insufficient_evidence has no citation nodes and no citations.
    SupplementaryAnswer:
      type: object
      required: [kind, blocks]
      properties:
        kind: {type: string, const: general_knowledge}
        blocks:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/SupplementaryBlock"
      additionalProperties: false
    AnswerRequest:
      type: object
      required: [query, mode]
      properties:
        query: {type: string, minLength: 1, maxLength: 1000}
        mode:
          $ref: "#/components/schemas/AnswerMode"
        filters:
          $ref: "#/components/schemas/SearchFilters"
        top_k: {type: integer, minimum: 1, maximum: 50, default: 10}
        response_length:
          allOf:
            - $ref: "#/components/schemas/ResponseLength"
          default: medium
      additionalProperties: false
    AnswerResponse:
      type: object
      required: [answer_id, query, mode, textbook, supplementary, request_id, created_at]
      properties:
        answer_id: {type: string, format: uuid}
        query: {type: string}
        mode:
          $ref: "#/components/schemas/AnswerMode"
        textbook:
          $ref: "#/components/schemas/TextbookAnswer"
        supplementary:
          oneOf:
            - $ref: "#/components/schemas/SupplementaryAnswer"
            - type: "null"
        request_id:
          $ref: "#/components/schemas/RequestIdentifier"
        created_at: {type: string, format: date-time}
      additionalProperties: false
...
    AnswerStartedEvent:
      type: object
      required: [answer_id, request_id]
      properties:
        answer_id: {type: string, format: uuid}
        request_id:
          $ref: "#/components/schemas/RequestIdentifier"
      additionalProperties: false
    AnswerProgressEvent:
      type: object
      required: [answer_id, stage]
      properties:
        answer_id: {type: string, format: uuid}
        stage:
          type: string
          enum: [retrieval, augmentation, generation, validation]
      additionalProperties: false
    AnswerDeltaEvent:
      type: object
      required: [answer_id, section, content, provisional]
      properties:
        answer_id: {type: string, format: uuid}
        section:
          type: string
          enum: [textbook, supplementary]
        content: {type: string}
        provisional: {type: boolean, const: true}
      additionalProperties: false
    AnswerResetEvent:
      type: object
      required: [answer_id, reason]
      properties:
        answer_id: {type: string, format: uuid}
        reason:
          type: string
          const: generation_restarted
      additionalProperties: false